Added support for Magento 2

+++ build.sh
@@ -27,21 +27,22 @@ git init
 git remote add origin $REPO_REMOTE
 cd ../
 
-# Parse Magento's site for a list of current releases
-wget --no-check-certificate -qO - http://www.magentocommerce.com/download | ack-grep '>magento-([0-9.]+)\.zip ' --output='$1' | sort -n > versions
-
 function download_release()
 {
-  local VERS="$1"
+  local VERS FORMAT FILENAME
 
-  if [ -f "$DOWNLOADS_DIR/magento-$VERS.zip" ]; then
-    cp $DOWNLOADS_DIR/magento-$VERS.zip magento-$VERS.zip
+  VERS="$1"
+  FORMAT="$2"
+  FILENAME=$(printf "$FORMAT" "$VERS")
+
+  if [ -f "$DOWNLOADS_DIR/$FILENAME" ]; then
+    cp $DOWNLOADS_DIR/$FILENAME magento-$VERS.zip
     return $?
   fi
 
-  wget --no-check-certificate -S --spider "$ASSET_URL/magento-$VERS.zip" 2>&1 | grep -q "Remote file exists"
+  wget --no-check-certificate -S --spider "$ASSET_URL/$FILENAME" 2>&1 | grep -q "Remote file exists"
   if [ $? -eq 0 ]; then
-    wget --no-check-certificate -qO magento-$VERS.zip $ASSET_URL/magento-$VERS.zip
+    wget --no-check-certificate -qO magento-$VERS.zip $ASSET_URL/$FILENAME
     return 0
   fi
 
@@ -55,6 +56,21 @@ function process_release()
   VERS="$1"
   COUNT="$2"
 
+  case ${VERS:0:1} in
+    1)
+      FORMAT="magento-%s.zip"
+      ;;
+    2)
+      FORMAT="Magento-CE-%s.zip"
+      ;;
+  esac
+
+  # Check to see if tag alreayd exists for release
+  if echo "$REMOTE_TAGS" | grep -qE "^$VERS\$"; then
+    echo -e "\n >> Skipping $VERS"
+    continue
+  fi
+
   # This is to work around the bug of some Magento downloads not being inside a directory called magento
   cd $START_DIR/magento-ce-builder
   [ -d "magento" ] && rm -rf magento
@@ -62,7 +78,7 @@ function process_release()
   cd magento
 
   echo -ne "\n >> Downloading $VERS"
-  download_release $VERS
+  download_release "$VERS" "$FORMAT"
   echo " OK"
 
   echo -n " >> Extracting"
@@ -84,7 +100,7 @@ function process_release()
   fi
   rsync -a --delete ../magento/ ../repo/ --exclude="/.git"
   git add * .htaccess* >/dev/null 2>&1
-  git commit -am "Version $VERS -new" >/dev/null 2>&1
+  git commit -am "Version $VERS" >/dev/null 2>&1
   git tag $VERS >/dev/null 2>&1
   echo " OK"
 
@@ -97,6 +113,10 @@ function process_release()
   return 0
 }
 
+# Parse Magento's site for a list of current releases
+wget --no-check-certificate -qO - http://www.magentocommerce.com/download | ack-grep '>(magento\-|Magento Community Edition )([0-9.]+)\.zip ' --output='$2' | sort -n > versions
+
+REMOTE_TAGS=$( cd $START_DIR/magento-ce-builder/repo; git ls-remote --tags origin | ack-grep 'tags/([0-9.]+)' --output='$1' )
 COUNT=0
 while read VERS; do
   process_release $VERS $COUNT
